test(research): extract source completion fixtures
